+++ root/app.js
@@ -1,4 +1,3 @@
-const calc = require('postcss-calc')
 const autoprefixer = require('autoprefixer')
 const webpack = require('webpack')
 const htmlStandards = require('reshape-standard')
@@ -9,7 +8,6 @@ const env = process.env.SPIKE_ENV
 
 function css (options) {
   const plugins = [
-    calc(options.calc),
     autoprefixer(options.autoprefixer)
   ]
 
@@ -25,27 +23,6 @@ function locals (ctx) {
     foo: 'bar'
   }
 }
-
-// Bootstrap 4 Browser Support
-const browsers = [
-  // Official browser support policy:
-  // https://v4-alpha.getbootstrap.com/getting-started/browsers-devices/#supported-browsers
-  //
-  'Chrome >= 45', // Exact version number here is kinda arbitrary
-  'Firefox ESR',
-  // Note: Edge versions in Autoprefixer & Can I Use refer to the EdgeHTML rendering engine version,
-  // NOT the Edge app version shown in Edge's "About" screen.
-  // For example, at the time of writing, Edge 20 on an up-to-date system uses EdgeHTML 12.
-  // See also https://github.com/Fyrd/caniuse/issues/1928
-  'Edge >= 12',
-  'Explorer >= 10',
-  // Out of leniency, we prefix these 1 version further back than the official policy.
-  'iOS >= 9',
-  'Safari >= 9',
-  // The following remain NOT officially supported, but we're lenient and include their prefixes to avoid severely breaking in them.
-  'Android >= 4.4',
-  'Opera >= 30'
-]
 
 module.exports = {
   devtool: 'source-map',
@@ -78,8 +55,7 @@ module.exports = {
     minify: env === 'production'
   }),
   postcss: css({
-    minify: env === 'production',
-    autoprefixer: { browsers: browsers }
+    minify: env === 'production'
   }),
   babel: jsStandards()
 }