+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.3.0] - 2017-10-05
+### Changed
+- Remove postcss-calc because bootstrap css no longer need it
+- Move browser list from app.js to package.json
+
 ## [0.2.0] - 2017-08-23
 ### Changed
 - Use reshape-standard v3.0